use Float decimal to calculate 68th percentile confidence level accuracyMeters instead of using Double and converting to Float after

+++ src/app/grapheneos/networklocation/NetworkLocationRepository.kt
@@ -45,7 +45,7 @@ class NetworkLocationRepository(
                         referenceDistance * 10f.pow((pathLoss - pathLossAtReferenceDistance) / (10f * pathLossExponent))
                     }
                     // should be at the 68th percentile confidence level
-                    (firstNearbyWifiAccessPoint.positioningData.accuracyMeters * 0.32).toFloat() + distanceFromAccessPoint
+                    (firstNearbyWifiAccessPoint.positioningData.accuracyMeters * 0.32f) + distanceFromAccessPoint
                 }
             }
             location
